+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Version 0.0.7
 
+- New: Highlight selected jump target token [#1](https://github.com/skafau/vscode-cursor-commander/issues/1)
 - New: Implemented jump cmds:
-  - Select to
+  - Select to [#4](https://github.com/skafau/vscode-cursor-commander/issues/4)
 
 [Changes](https://github.com/skafau/vscode-cursor-commander/compare/v0.0.6...v0.0.7)