+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 0.0.8
+
+- Change: Improve performance of token highlighting
+
+[Changes](https://github.com/skafau/vscode-cursor-commander/compare/v0.0.7...v0.0.8)
+
 ## Version 0.0.7
 
 - New: Highlight selected jump target token [#1](https://github.com/skafau/vscode-cursor-commander/issues/1)